requirements.txt
torch>=1.5.1 
torchvision>=0.6.1
matplotlib
numpy
argparse==1.4.0
tqdm
pyyaml